requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
cx-Freeze==6.15.7
discord.py==2.3.2
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
outcome==1.2.0
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
supervisor==4.2.5
trio==0.22.2
trio-websocket==0.10.4
tzlocal==5.0.1
urllib3==1.26.16
urllib3-secure-extra==0.1.0
wsproto==1.2.0
yarl==1.9.2
